rename(ResultCommandDisplay): 사용하지 않는 display enum 제거
